MDL-49227 mod_forum: Improve discussion subscription on post
If the user has posted in the forum before, follow their previous
discussion preference.

If the user has not posted before, but is subscribed to the forum, follow
their forum preference.

Otherwise, follow their user-preference.

This has the result that if a user has unsubscribed from a forum and writes
their first response; the user preference will be followed. This may
override the forum subscription.

This commit also changes from a dropdown to a checkbox in the post form.

This is a backport of MDL-48221